[bazel] Switch to custom `closure_js_deps` rule (#16571)
After `rules_closure` 0.12.0, the `closure_js_deps` rule was removed
and it was advised to use an `npm` invocation to generate the deps
file. However, we still want to use the deps file as it makes it a
lot easier to iterate on the atoms when we work on them. The obvious
solution is to write our own implementation, which is what happens
in this commit.