Avoid sleep in tests (#16713)
* remove useless sleeps from SessionHandlingTest

This change decrease execution time of `SessionHandlingTest` from ~28s to 7s.

* Avoid `sleep(2000)` in hover tests

Instead of wasting 2 second (which doesn't guarantee the result), just wait for the needed element that should appear as a result of hovering.

* use much shorter polling interval to speed up the tests

if the first try failed, wasting the whole 0.5 second is a huge waste of time. Let's try sooner!

* Avoid `sleep` in drag'n'drop tests

fix the test page usability: make the elements visible only after they are fully initialized.

To force revealing potential flaky tests in the future, I've added a small pause for the initialization code. So the tests should wait until the element appears on the screen - then it's ready for manipulations.

* Open a blank page before every test

It's a good practice to avoid tests affecting each other:
1. Open "about:blank" page - this stops any current activity / background requests / animations on the previous page
2. Open an empty page which clears sessionStorage, localStorage and cookies.

This technique allows reusing the browser between tests, while keeping the tests independent.

* tiny cleanup in ExpectedConditions.java

* fix potential NPE in tearDown methods

Also, reset the link to static field after test run - to avoid holding heavy objects in memory till the end of all tests.