[java] [refactor] replace most of Guava collections by Java built-in collections (#16776)
* [refactor] replace `Arrays.asList` by `List.of`

`List.of` is preferred because it
1. Optimized, and
2. Returns a truly immutable object.

While `Arrays.asList` returns a list that allows modifications, e.g. using method `*.set(index, value)`.

* [refactor] replace `ImmutableMap.of` by `Map.of`

Guava's method `ImmutableMap.of` was needed until in Java 8 edge, when Java didn't have its own constructor for Maps. Nowadays `Map.of` is the proper way to create maps.

* [refactor] replace `ImmutableList.of` by `List.of`

Guava's method `ImmutableList.of` was needed until in Java 8 edge, when Java didn't have its own constructor for Lists. Nowadays `List.of` is the proper way to create immutable lists.

* [refactor] replace `ImmutableSet.of` by `Set.of`

Guava's method `ImmutableSet.of` was needed until in Java 8 edge, when Java didn't have its own constructor for Sets. Nowadays `Set.of` is the proper way to create immutable Sets.

Sometimes we still need to use `ImmutableSet` - where we need to keep the order of elements.

* use single imports instead of `import x.y.*`